Luke: (DataStore) Refactored DM_Datastore into DataStore
- Huge refactor
- Moved all necessary functionality out of DM_DataStore into DataStore
- Added query_view and utility methods to DataStore
- moved sha1hex into couhdb_datastore.py
- Deleted couchdb_dm_datastore.py
