+++ pyon/datastore/couchdb/couchdb_config.py
@@ -25,6 +25,12 @@ COUCHDB_CONFIGS = {
     DataStore.DS_PROFILE.SCIDATA:{
         'views': []
     },
+    DataStore.DS_PROFILE.DM: {
+        'views':[]
+    },
+    DataStore.DS_PROFILE.EXAMPLES:{
+        'views':['posts']
+    },
     DataStore.DS_PROFILE.BASIC:{
         'views': []
     },
@@ -196,6 +202,35 @@ function(doc) {
 }""",
         },
     },
+    'posts' : {
+        "posts_by_id": {
+            "map": "function(doc)\n{\tif(doc.type_==\"BlogPost\") { emit(doc.post_id,doc._id);}}"
+        },
+        "posts_by_title": {
+            "map": "function(doc)\n{\tif(doc.type_==\"BlogPost\") { emit(doc.title,doc._id);}}"
+        },
+        "posts_by_updated": {
+            "map": "function(doc)\n{\tif(doc.type_==\"BlogPost\") { emit(doc.updated,doc._id);}}"
+        },
+        "posts_by_author": {
+            "map": "function(doc)\n{\tif(doc.type_==\"BlogPost\") { emit(doc.author.name,doc._id);}}"
+        },
+        "comments_by_post_id": {
+            "map": "function(doc)\n{\tif(doc.type_==\"BlogComment\") { emit(doc.ref_id,doc._id);}}"
+        },
+        "comments_by_author": {
+            "map": "function(doc)\n{\tif(doc.type_==\"BlogComment\") { emit(doc.author.name,doc._id);}}"
+        },
+        "comments_by_updated": {
+            "map": "function(doc)\n{\tif(doc.type_==\"BlogComment\") { emit(doc.updated,doc._id);}}"
+        },
+        "posts_join_comments": {
+            "map": "function(doc)\n{\tif(doc.type_==\"BlogPost\") { emit([doc.post_id,0],doc._id);}\n\telse if(doc.type_==\"BlogComment\") { emit([doc.ref_id,1],doc._id);}\n}"
+        },
+        "posts_by_author_date": {
+            "map": "function(doc) {\n  if(doc.type_==\"BlogPost\")\n    emit([doc.author.name,doc.updated,doc.post_id], doc.post_id);\n  else if(doc.type==\"BlogComment\")\n    emit([doc.author.name,doc.updated,doc.ref_id], doc.ref_id);\n}"
+        }
+    }
 }
 
 def get_couchdb_views(config):
